+++ app/rdev/framework/console/commands/FlushFrameworkCacheCommand.php
@@ -5,7 +5,7 @@
  * Defines the command that flushes the framework's cache
  */
 namespace RDev\Framework\Console\Commands;
-use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
+use RDev\Applications\Bootstrappers\IO\IBootstrapperIO;
 use RDev\Applications\Paths;
 use RDev\Console\Commands\Command;
 use RDev\Console\Responses\IResponse;
@@ -14,14 +14,14 @@ class FlushFrameworkCacheCommand extends Command
 {
     /** @var Paths The application paths */
     private $paths = null;
-    /** @var BootstrapperIO The bootstrapper IO */
+    /** @var IBootstrapperIO The bootstrapper IO */
     private $bootstrapperIO = null;
 
     /**
      * @param Paths $paths The application paths
-     * @param BootstrapperIO $bootstrapperIO The bootstrapper IO
+     * @param IBootstrapperIO $bootstrapperIO The bootstrapper IO
      */
-    public function __construct(Paths $paths, BootstrapperIO $bootstrapperIO)
+    public function __construct(Paths $paths, IBootstrapperIO $bootstrapperIO)
     {
         parent::__construct();
 
@@ -57,12 +57,12 @@ class FlushFrameworkCacheCommand extends Command
 
         if(isset($this->paths["tmp.framework.console"]))
         {
-            $fileNames[] = $this->paths["tmp.framework.console"] . "/" . BootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME;
+            $fileNames[] = $this->paths["tmp.framework.console"] . "/" . IBootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME;
         }
 
         if(isset($this->paths["tmp.framework.http"]))
         {
-            $fileNames[] = $this->paths["tmp.framework.http"] . "/" . BootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME;
+            $fileNames[] = $this->paths["tmp.framework.http"] . "/" . IBootstrapperIO::DEFAULT_CACHED_REGISTRY_FILE_NAME;
         }
 
         foreach($fileNames as $cachedRegistryFileName)