+++ app/rdev/framework/tests/ApplicationTestCase.php
@@ -6,6 +6,7 @@
  */
 namespace RDev\Framework\Tests;
 use PHPUnit_Framework_TestCase;
+use Monolog\Logger;
 use RDev\Applications\Application;
 
 abstract class ApplicationTestCase extends PHPUnit_Framework_TestCase
@@ -28,6 +29,13 @@ abstract class ApplicationTestCase extends PHPUnit_Framework_TestCase
     {
         $this->application->shutdown();
     }
+
+    /**
+     * Gets the kernel logger
+     *
+     * @return Logger The logger to use in the kernel
+     */
+    abstract protected function getKernelLogger();
 
     /**
      * Sets the instance of the application to use in tests