+++ tests/app/rdev/applications/bootstrappers/io/BootstrapperIOTest.php
@@ -35,7 +35,7 @@ class BootstrapperIOTest extends \PHPUnit_Framework_TestCase
             "tmp.framework" => __DIR__ . "/files"
         ]);
         $this->environment = new Environment(Environment::TESTING);
-        $this->io = new BootstrapperIO($this->paths, $this->environment);
+        $this->io = new BootstrapperIO($this->paths);
         $this->registry = new BootstrapperRegistry($this->paths, $this->environment);
     }