+++ tests/app/rdev/framework/tests/http/mocks/ApplicationTestCase.php
@@ -11,8 +11,11 @@ use RDev\Applications\Bootstrappers\BootstrapperRegistry;
 use RDev\Applications\Bootstrappers\Dispatchers\Dispatcher;
 use RDev\Applications\Environments\Environment;
 use RDev\Applications\Paths;
+use RDev\Applications\Tasks\Dispatchers\Dispatcher as TaskDispatcher;
+use RDev\Applications\Tasks\TaskTypes;
 use RDev\Framework\Tests\HTTP\ApplicationTestCase as BaseApplicationTestCase;
 use RDev\IoC\Container;
+use RDev\IoC\IContainer;
 use RDev\Tests\Applications\Mocks\MonologHandler;
 
 class ApplicationTestCase extends BaseApplicationTestCase
@@ -35,27 +38,37 @@ class ApplicationTestCase extends BaseApplicationTestCase
     /**
      * {@inheritdoc}
      */
+    protected function getKernelLogger()
+    {
+        $logger = new Logger("application");
+        $logger->pushHandler(new MonologHandler());
+
+        return $logger;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     protected function setApplication()
     {
         // Create and bind all of the components of our application
         $paths = new Paths([
             "configs" => __DIR__ . "/../../configs"
         ]);
-        $logger = new Logger("application");
-        $logger->pushHandler(new MonologHandler());
+        $taskDispatcher = new TaskDispatcher();
         $environment = new Environment(Environment::TESTING);
         $container = new Container();
-        $container->bind("RDev\\Applications\\Paths", $paths);
-        $container->bind("Monolog\\Logger", $logger);
-        $container->bind("RDev\\Applications\\Environments\\Environment", $environment);
-        $container->bind("RDev\\IoC\\IContainer", $container);
-        $this->application = new Application($paths, $logger, $environment, $container);
+        $container->bind(Paths::class, $paths);
+        $container->bind(TaskDispatcher::class, $taskDispatcher);
+        $container->bind(Environment::class, $environment);
+        $container->bind(IContainer::class, $container);
+        $this->application = new Application($paths, $taskDispatcher, $environment, $container);
 
         // Setup the bootstrappers
         $bootstrapperRegistry = new BootstrapperRegistry($paths, $environment);
-        $bootstrapperDispatcher = new Dispatcher($this->application);
+        $bootstrapperDispatcher = new Dispatcher($taskDispatcher, $container);
         $bootstrapperRegistry->registerEagerBootstrapper(self::$bootstrappers);
-        $this->application->registerPreStartTask(function () use ($bootstrapperDispatcher, $bootstrapperRegistry)
+        $taskDispatcher->registerTask(TaskTypes::PRE_START, function () use ($bootstrapperDispatcher, $bootstrapperRegistry)
         {
             $bootstrapperDispatcher->dispatch($bootstrapperRegistry);
         });